fixup! add wrapper to functions
