+++ backend/python/app/__init__.py
@@ -66,7 +66,10 @@ def create_app(config_name):
     )
 
     app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
-    app.config["SQLALCHEMY_ENGINE_OPTIONS"] = {"pool_recycle": 1800}
+    app.config["SQLALCHEMY_ENGINE_OPTIONS"] = {
+        "pool_pre_ping": True,
+        "pool_recycle": 1800,
+    }
     mysql.init_app(app)
     firebase_admin.initialize_app()
