+++ backend/python/app/services/implementations/language_service.py
@@ -4,35 +4,13 @@ from sqlalchemy import exc
 from ...models import db
 from ...models.language import Language
 from ..interfaces.language_service import ILanguageService
-
-
-def handle_exceptions(f):
-    from functools import wraps
-
-    @wraps(f)
-    def wrapper(*args, **kwargs):
-        try:
-            res = f(*args, **kwargs)
-            db.session.commit()
-            return res
-        except exc.SQLAlchemyError as error:
-            db.session.rollback()
-            raise error
-        except Exception as error:
-            self = args[0]
-            self.logger.error(error)
-            raise error
-        finally:
-            db.session.close()
-
-    return wrapper
+from .utils import handle_exceptions
 
 
 class LanguageService(ILanguageService):
     def __init__(self, logger=current_app.logger):
         self.logger = logger
 
-    @handle_exceptions
     def get_languages(self):
         try:
             languages = Language.query.order_by(Language.language.asc()).all()
@@ -52,7 +30,6 @@ class LanguageService(ILanguageService):
             self.logger.error(str(error))
             raise error
 
-    @handle_exceptions
     def get_is_rtl(self, language):
         try:
             to_check = Language.query.filter_by(language=language).first()
